Nit: consistent referring to company VMware
Officially, vmware is written as "VMware", not "VMWare" or other
variations. This is a common mistake.

I also see inconsistency throughout the codebase on written form
of vmware. This change updates the UI facing messages to be
consisstent at least.

It might be benefitual to have a following PR that moves the "VMware"
string into a string substitution variable for consistent referencing.

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>