More robust handling of vmmetadataexporter CLI args
This change does more validation of the CLI args given to
vmmetadataexporter to avoid panics. If host isn't given, it
errors and exits. If username or password isn't given, it
prompts the user to input them.

This handles a variety of scenarios around giving or not giving
these three arguments. It is also a better solution around
handling a password, since a CLI password arg gets leaked to
the shell history and potentially other logs.

Fixes #165

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>