Fix panic caused by OpenstackCreds CA cert check
Currently, the check passes even when there is a failure to get
certificate from the endpoint. This is due to a typo in error check,
we end up checking for wrong err variable.

Because of it ends up with a nil pointer dereference panic in
GetOpenstackInfo when it is called by migration controller to get
details of volume etc.

Test plan:
1. Built a new controller image.
2. Uploaded a OpenStack RC file with HTTP auth URL and without
   OS_INSECURE.
3. The validation falied as expected.

Fixes https://github.com/platform9/vjailbreak/issues/234

Signed-off-by: Bhavin Gandhi <bhavin7392@gmail.com>